# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Olist Checkout API",
        "description": "API REST para sistema de checkout da Olist",
        "contact": {
            "name": "Olist Checkout",
            "email": "support@olist.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Listar pedidos do usuário",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pedidos paginada"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Criar pedido",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items",
                                    "shipping_address"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "shipping_address": {
                                        "required": [
                                            "street",
                                            "city",
                                            "state",
                                            "zip_code",
                                            "country"
                                        ],
                                        "properties": {
                                            "street": {
                                                "type": "string",
                                                "example": "Rua Exemplo, 123"
                                            },
                                            "city": {
                                                "type": "string",
                                                "example": "São Paulo"
                                            },
                                            "state": {
                                                "type": "string",
                                                "example": "SP"
                                            },
                                            "zip_code": {
                                                "type": "string",
                                                "example": "01234-567"
                                            },
                                            "country": {
                                                "type": "string",
                                                "example": "Brasil"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido criado com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação ou estoque insuficiente"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Obter pedido",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do pedido"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Atualizar pedido",
                "operationId": "6a7029f5510ca08e28a2d58a5fe19439",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "processing",
                                            "shipped",
                                            "delivered",
                                            "cancelled"
                                        ],
                                        "example": "processing"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido atualizado"
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}